Update the response of REST api
The original version of the api returns a dict representation of the model object (Medicine.asdict()). On the frontend side we can't use the fetch api, because the input of self.wfile.write() is not a valid JSON object (eg.: it contains apostrophe instead of quotes). 
about the json lib: https://www.w3schools.com/python/python_json.asp
After this change we can call the REST api with fetch.
Example:
1. Run server
  python server.py 3000 localhost
2. Fetch Rest api
  fetch("http://localhost:3000/api/getAll").then(response => response.json()).then(list=>console.log(list))